requirements.txt
streamlit
tensorflow
numpy
pillow
opencv-python-headless
scikit-learn
matplotlib
pyrebase4
firebase-admin